+++ src/app.jsx
@@ -1,10 +1,17 @@
 import React, { useEffect } from 'react';
-import './app.css';
+import './App.css';
 import { useSelector, useDispatch } from 'react-redux';
-import { Routes, Route, useNavigate } from 'react-router-dom';
+import {
+  Routes, Route, useNavigate, Navigate,
+} from 'react-router-dom';
 import { checkLogin } from './redux/user/userSlice';
-import Dashboard from './components/Dashboard';
-import LandingPage from './components/LandingPage';
+import Dashboard from './components/pages/Home/Dashboard';
+import LandingPage from './components/pages/sessions/LandingPage';
+import ReserveForm from './components/pages/Reservations/ReserveForm';
+import MyReservations from './components/pages/Reservations/MyReservations';
+import AddCarItem from './components/pages/Car/AddCarItem';
+import DeleteCarItem from './components/pages/Car/DeleteCarItem';
+import AppLayout from './components/shared/AppLayout';
 
 const App = () => {
   const { loggedIn } = useSelector((state) => state.user);
@@ -24,13 +31,21 @@ const App = () => {
   }, [loggedIn]);
 
   return (
-
-    <>
+    <div>
       <Routes>
-        <Route path="/" element={<LandingPage />} />
-        <Route path="/dashboard" element={<Dashboard />} />
+        <Route index element={<Navigate replace to="login" />} />
+        <Route path="/login" element={<LandingPage />} />
+        <Route path="/register" element={<LandingPage />} />
+        <Route element={<AppLayout />}>
+          <Route path="/dashboard" element={<Dashboard />} />
+          <Route path="/reserve-form" element={<ReserveForm />} />
+          <Route path="/my-reservations" element={<MyReservations />} />
+          <Route path="/add-car-item" element={<AddCarItem />} />
+          <Route path="/delete-car-item" element={<DeleteCarItem />} />
+        </Route>
       </Routes>
-    </>
+    </div>
   );
 };
+
 export default App;
