+++ src/components/pages/sessions/LandingPage.jsx
@@ -1,19 +1,34 @@
 import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
 import Login from './Login';
 import Registrations from './Registrations';
 
 const LandingPage = () => {
   const [login, setLogin] = useState(true);
   const { loading } = useSelector((state) => state.user);
+  const history = useNavigate();
+
   if (loading) {
     return (<h1 className="loading">{loading ? 'Loading...' : ''}</h1>);
   }
   return (
     <>
-
       {login ? <Login /> : <Registrations />}
-      <button type="button" className="toggleLogin" onClick={() => setLogin(!login)}>{login ? 'Register' : 'Login'}</button>
+      <button
+        type="button"
+        className="toggleLogin"
+        onClick={() => {
+          setLogin(!login);
+          if (login) {
+            history('/register');
+          } else {
+            history('/login');
+          }
+        }}
+      >
+        {login ? 'Register' : 'Login'}
+      </button>
     </>
   );
 };
